Fix max screen pages like the homepage
The new rules added in global.css ensure that page content always takes
up the entire vertical space if the viewport exceeds the content. This
will push the footer down to the bottom of the page consistently.
With this improved layout, we can reposition the PotionBackground
component within `<main>` and use absolute positioning to fill the
background for `<main>`.

With this layout, we can remove hard-coded z-indexing to ensure the
PotionBackground appears behind other elements and instead we give it
a negative z-index for certainty.